// frontend/src/app/core/api/model/grade-record-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GradeRecordResponse { 
    id?: number;
    enrollmentId?: number;
    matricule?: string;
    nom?: string;
    prenom?: string;
    presence?: GradeRecordResponse.PresenceEnum;
    note?: number;
    commentaire?: string;
}
export namespace GradeRecordResponse {
    export const PresenceEnum = {
        Present: 'PRESENT',
        AbsentJustified: 'ABSENT_JUSTIFIED',
        AbsentUnjustified: 'ABSENT_UNJUSTIFIED'
    } as const;
    export type PresenceEnum = typeof PresenceEnum[keyof typeof PresenceEnum];
}